Use a sequence to generate the id and create an id object in service.

+++ src/main/java/de/intevation/lada/model/LMesswertId.java
@@ -12,13 +12,13 @@ import javax.persistence.Embeddable;
 public class LMesswertId implements java.io.Serializable {
 
 	private String probeId;
-	private short messungsId;
+	private Integer messungsId;
 	private int messgroesseId;
 
 	public LMesswertId() {
 	}
 
-	public LMesswertId(String probeId, short messungsId, int messgroesseId) {
+	public LMesswertId(String probeId, Integer messungsId, int messgroesseId) {
 		this.probeId = probeId;
 		this.messungsId = messungsId;
 		this.messgroesseId = messgroesseId;
@@ -34,11 +34,11 @@ public class LMesswertId implements java.io.Serializable {
 	}
 
 	@Column(name = "messungs_id", nullable = false)
-	public short getMessungsId() {
+	public Integer getMessungsId() {
 		return this.messungsId;
 	}
 
-	public void setMessungsId(short messungsId) {
+	public void setMessungsId(Integer messungsId) {
 		this.messungsId = messungsId;
 	}
